gen_coindesk_api_calls complete - now creates an arbitrry number of *.rng files with api calls for each day's data

+++ data/aquire/timestamp.py
@@ -5,7 +5,7 @@ class TimeStamp:
     def totimestamp(self,dt, epoch=datetime(1970,1,1)):
         td = dt - epoch
         # return td.total_seconds()
-        return (td.microseconds + (td.seconds + td.days * 86400) * 10**6) / 10**6 
+        return int((td.microseconds + (td.seconds + td.days * 86400) * 10**6) / 10**6) 
     def todate(self,timestamp):
         d = datetime.fromtimestamp(timestamp)
         return d
